Unify navigation state rendering
Some missing capabilites are added for each renderer type
* Error handling now supports postponing and rerouting
* Status code set through the event for regular navigation is now
validated

Fixes #3648
